pyproject.toml
[tool.poetry]
name = "cover-lettered"
version = "0.1.0"
description = ""
authors = ["Joe Foster <joe.foster@protonmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "=3.0.0"
flask-sqlalchemy = "^3.1.1"
wikipedia = "^1.4.0"
rich = "^13.9.3"
pathlib = "^1.0.1"
requests = "^2.32.3"
pyyaml = "^6.0.2"
transformers = "^4.46.1"
tensorflow = "^2.18.0"
openai = "^1.54.2"
python-docx = "^1.1.2"
pypandoc = "^1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
